Remove (invalid) byte order mark (BOM) from targetver.h
This file started with the two bytes 0xFF 0xFE. This is the byte order
mark (U+FEFF) in Little-Endian, which is the default on Windows. However,
it is not valid UTF-8. To be UTF-8 the code points has to be encoded using
big-endian, resulting in the BOM to be stored as 0xFE 0xFF.

This commit simply strips the first two bytes, since the BOM is not
needed.

This issue was discovered thanks to unicop:
https://github.com/mullvad/unicop